# nodejs-server-server-generated/api/openapi.yaml
openapi: 3.0.0
info:
  title: Incident Management API
  description: "API para la gestión de incidentes, partes de seguros y evidencias."
  version: 1.0.0
servers:
- url: http://api.incident-management.com/v1
  description: Servidor de producción
...
paths:
...
  /notificar/:
    put:
      tags:
      - Notificar
      description: Notifica al usuario la resolucion
      operationId: notificarId_usuarioPUT
      parameters:
      - name: dni_usuario
        in: query
        required: true
        schema:
          type: string
      - name: aprobado
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: Notificado correctamente
        "400":
          description: Datos inválidos
      x-swagger-router-controller: Notificar
...